pyproject.toml
[project]
name = "kegg-pathways-completeness"
version = "1.2.1"
readme = "README.md"
license = {text = "Apache Software License 2.0"}
authors = [
  { name = "Ekaterina Sakharova", email = "kates@ebi.ac.uk" },
]
keywords = ["bioinformatics", "pipelines", "metagenomics", "kegg"]
description = "The tool counts completeness of each KEGG pathway for protein sequences."
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "biopython>=1.83",
    "networkx>=3.3",
    "graphviz>=0.20.3",
    "pydot>=3.0.4",
    "tomli>=2.0.1"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "kegg_pathways_completeness",
    "kegg_pathways_completeness.bin", 
    "kegg_pathways_completeness.bin.parse_hmmtable",
    "kegg_pathways_completeness.pathways_data"
]

[tool.setuptools.package-data]
"*" = ["graphs.pkl", "*.txt"]

[project.scripts]
give_completeness = "kegg_pathways_completeness.bin.give_completeness:main"
plot_modules_graphs = "kegg_pathways_completeness.bin.plot_modules_graphs:main"
make_graphs = "kegg_pathways_completeness.bin.make_graphs:main"
fetch_modules_data = "kegg_pathways_completeness.bin.fetch_modules_data:main"
make_hmmer_table_tab_separated = "kegg_pathways_completeness.bin.parse_hmmtable.make_hmmer_table_tab_separated:main"
parse_hmmer_tab_separated_table = "kegg_pathways_completeness.bin.parse_hmmtable.parse_hmmer_tab_separated_table:main"


[project.optional-dependencies]
dev = [
    "pre-commit==3.6.0",
    "black==24.4.2",
]
test = [
    "pytest==8.2.2",
    "pytest-workflow==2.1.0",
    "pytest-md==0.2.0",
]

[tool.pytest.ini_options]
testpaths = "tests"
required_plugins = "pytest-workflow"
addopts= "--git-aware"